Add files via upload
The initial DESO ontology. This is slightly dated - for example the lack of the exFat file system - but it is designed as a proof of concept. The location for 'Drone' is included to show how the Location class can be expanded and manipulated to keep track of technological chances. For background on this ontology, please see https://kclpure.kcl.ac.uk/portal/files/95696488/2018_Brady_Owen_0966104_ethesis.pdf